+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 14.25.0 (2017-04-20)
+
+* **Added** support for select renderer to be driven by an array of integers/numbers/strings when using `endpoint` or simply `recordsPath` to mine the form value.
+
+
+# 14.24.0 (2017-04-20)
+
+* **Added** support for mining select lists from the form value instead of an API endpoint.
+
+
 # 14.23.2 (2017-04-07)
 
 * **Fixed** boolean renderer to put label to right of checkbox.